Removed --disable-bootrstrap
GCC stage 2: removed one more archive retrieval

+++ scripts/005-gcc-8.2.0-stage2.sh
@@ -6,27 +6,33 @@
 GCC_VERSION=8.2.0
 ## Download the source code.
 SOURCE=http://ftpmirror.gnu.org/gcc/gcc-$GCC_VERSION/gcc-$GCC_VERSION.tar.gz
-wget --continue $SOURCE || { exit 1; }
 
-## Unpack the source code.
-echo Decompressing GCC $GCC_VERSION. Please wait.
-rm -Rf gcc-$GCC_VERSION && pigz -dc gcc-$GCC_VERSION.tar.gz | pv -ptrI | tar xf - || { exit 1; }
+if [ ! -d gcc-$GCC_VERSION ]; then
+	wget --continue $SOURCE || { exit 1; }
 
-## Enter the source directory and patch the source code.
-cd gcc-$GCC_VERSION || { exit 1; }
+	## Unpack the source code.
+	echo Decompressing GCC $GCC_VERSION. Please wait.
+	rm -Rf gcc-$GCC_VERSION && pigz -dc gcc-$GCC_VERSION.tar.gz | pv -ptrI | tar xf - || { exit 1; }
 
-## Apply the uncommitted patches first (libgcc, short-loop bug, EE-core extensions, MMI)
-## TODO: Remove these lines, once the patches are submitted.
-cat ../../patches/gcc-$GCC_VERSION-libgcc.patch | patch -p1 || { exit 1; }
-cat ../../patches/gcc-$GCC_VERSION-ee.patch | patch -p1 || { exit 1; }
-#cat ../../patches/gcc-$GCC_VERSION-mmi.patch | patch -p1 || { exit 1; }
-if [ -e ../../patches/gcc-$GCC_VERSION-PS2.patch ]; then
-	cat ../../patches/gcc-$GCC_VERSION-PS2.patch | patch -p1 || { exit 1; }
+	## Enter the source directory and patch the source code.
+	cd gcc-$GCC_VERSION || { exit 1; }
+
+	## Apply the uncommitted patches first (libgcc, short-loop bug, EE-core extensions, MMI)
+	## TODO: Remove these lines, once the patches are submitted.
+	cat ../../patches/gcc-$GCC_VERSION-libgcc.patch | patch -p1 || { exit 1; }
+	cat ../../patches/gcc-$GCC_VERSION-ee.patch | patch -p1 || { exit 1; }
+	#cat ../../patches/gcc-$GCC_VERSION-mmi.patch | patch -p1 || { exit 1; }
+	if [ -e ../../patches/gcc-$GCC_VERSION-PS2.patch ]; then
+		cat ../../patches/gcc-$GCC_VERSION-PS2.patch | patch -p1 || { exit 1; }
+	fi
+
+	## isl 0.20 broke GCC compilation: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=86724
+	## Remove once a new version is released.
+	cat ../../patches/gcc-$GCC_VERSION-isl.patch | patch -p1 || { exit 1; }
+	cd ..
 fi
 
-## isl 0.20 broke GCC compilation: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=86724
-## Remove once a new version is released.
-cat ../../patches/gcc-$GCC_VERSION-isl.patch | patch -p1 || { exit 1; }
+cd gcc-$GCC_VERSION || { exit 1; }
 
 ## OS Windows doesn't properly work with multi-core processors
 OSVER=$(uname)
@@ -45,7 +51,7 @@ TARG_XTRA_OPTS="--with-float=hard --with-newlib"
 mkdir build-$TARG_NAME-stage2 && cd build-$TARG_NAME-stage2 || { exit 1; }
 
 ## Configure the build.
-../configure --quiet --prefix="$PS2DEV/$TARG_NAME" --target="$TARGET" --enable-languages="c,c++" --disable-nls --disable-shared --disable-libssp --disable-libmudflap --disable-threads --disable-libgomp --disable-libquadmath --disable-target-libiberty --disable-target-zlib --without-ppl --without-cloog --with-headers=no --disable-libada --disable-libatomic --disable-multilib --disable-bootstrap --with-headers="$PS2DEV/$TARG_NAME/$TARGET/include" $TARG_XTRA_OPTS || { exit 1; }
+../configure --quiet --prefix="$PS2DEV/$TARG_NAME" --target="$TARGET" --enable-languages="c,c++" --disable-nls --disable-shared --disable-libssp --disable-libmudflap --disable-threads --disable-libgomp --disable-libquadmath --disable-target-libiberty --disable-target-zlib --without-ppl --without-cloog --with-headers=no --disable-libada --disable-libatomic --disable-multilib --with-headers="$PS2DEV/$TARG_NAME/$TARGET/include" $TARG_XTRA_OPTS || { exit 1; }
 
 ## Compile and install.
 make --quiet -j $PROC_NR && make --quiet install || { exit 1; }
